4.21 bloglist expansion step 9

+++ part4/bloglist/controllers/blogs.js
@@ -42,6 +42,13 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  console.log(decodedToken.id, blog.user.id)
+  if (!decodedToken.id || blog.user.toString() !== decodedToken.id.toString()) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
